Use long for larger values in oneshot timer intervals

+++ java/src/com/rubyeventmachine/EmReactor.java
@@ -379,7 +379,7 @@ public class EmReactor {
 		}
 	}
 
-	public long installOneshotTimer (BigInteger milliseconds) {
+	public long installOneshotTimer (long milliseconds) {
 		long s = createBinding();
 		long deadline = new Date().getTime() + milliseconds;
